pyproject.toml
[project]
name = "enterprise-business-system"
version = "0.1.0"
description = "Enterprise-grade Multi-Crew AI Orchestration System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai",
    "crewai-tools",
    "langchain",
    "openai",
    "pydantic",
    "python-dotenv",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "isort",
    "mypy",
    "ruff",
    "pre-commit",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B"]
ignore = []

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=crews --cov=flows"
testpaths = [
    "tests",
]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
